+++ applications/vessel-history/src/types/index.ts
@@ -1,4 +1,4 @@
-import { Vessel } from '@globalfishingwatch/api-types'
+import { Vessel, Authorization } from '@globalfishingwatch/api-types'
 import { UrlDataviewInstance } from '@globalfishingwatch/dataviews-client'
 
 export type WorkspaceViewportParam = 'latitude' | 'longitude' | 'zoom'
@@ -59,6 +59,7 @@ export interface VesselFieldsHistory {
 
 export interface VesselWithHistory extends Vessel {
   history: VesselFieldsHistory
+  iuuStatus?: number
 }
 
 export enum VesselAPISource {
@@ -117,18 +118,12 @@ export interface RelationList {
   vesselOwnership: VesselOwnership[]
   vesselOperations: VesselOperation[]
 }
-
-export interface AuthorizationList {
-  source: string
-  startDate: string
-  endDate?: string
-}
-
 export interface TMTDetail {
   vesselMatchId: string
   valueList: ValueList
+  iuuStatus: number
   relationList: RelationList
-  authorisationList: AuthorizationList[]
+  authorisationList: Authorization[]
   imageList: string[]
 }