+++ applications/dataviews-editor/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @globalfishingwatchapp/dataviews-editor
 
+## 1.1.86
+
+### Patch Changes
+
+- Updated dependencies [5fccc9f4]
+- Updated dependencies [06bbf6b9]
+  - @globalfishingwatch/layer-composer@9.1.0
+
+## 1.1.85
+
+### Patch Changes
+
+- @globalfishingwatch/api-client@3.0.0
+- @globalfishingwatch/dataviews-client@12.0.0
+- @globalfishingwatch/layer-composer@9.0.0
+
 ## 1.1.84
 
 ### Patch Changes