+++ applications/tile-inspector/CHANGELOG.md
@@ -1,5 +1,23 @@
 # tile-inspector
 
+## 0.1.76
+
+### Patch Changes
+
+- Updated dependencies [5fccc9f4]
+- Updated dependencies [06bbf6b9]
+  - @globalfishingwatch/layer-composer@9.1.0
+  - @globalfishingwatch/ui-components@10.1.0
+
+## 0.1.75
+
+### Patch Changes
+
+- @globalfishingwatch/api-client@3.0.0
+- @globalfishingwatch/layer-composer@9.0.0
+- @globalfishingwatch/react-hooks@13.0.0
+- @globalfishingwatch/ui-components@10.0.0
+
 ## 0.1.74
 
 ### Patch Changes